parse_cmd_onewire_get: signed output of temperature.

+++ ecmd_parser/onewire.c
@@ -194,7 +194,7 @@ int16_t parse_cmd_onewire_get(char *cmd, char *output, uint16_t len)
 
         ret = snprintf_P(output, len,
                 PSTR("Temperatur: %3d.%1d"),
-                HI8(temp),  HI8(((temp & 0x00ff) * 10) + 0x80));
+                (int8_t) HI8(temp),  HI8(((temp & 0x00ff) * 10) + 0x80));
 
 #ifdef ONEWIRE_DS2502_SUPPORT
     } else if (ow_eeprom(&rom)) {
